requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.9
Pillow==10.4.0
opencv-python-headless==4.10.0.84
ultralytics==8.3.26
numpy==1.26.4
